+++ packages/map-template/src/i18n/nl.js
@@ -78,6 +78,9 @@ const nl = {
     // Opening Hours
     'Open': 'Open',
     'Closed': 'Gesloten',
+    // View Selector
+    'Pan map to view': 'Kaart verschuiven naar weergave',
+    'Close view selector': 'Sluiten weergavekiezer',
 };
 
 export default nl;